mm, oom: normalize oom scores to oom_score_adj scale only for userspace (bnc#884582, bnc#884767).
suse-commit: 8fb0b7d329ec3df8609e2f323c1f5b24b4bf127d

+++ include/linux/oom.h
@@ -49,7 +49,7 @@ enum oom_scan_t {
 
 extern int test_set_oom_score_adj(int new_val);
 
-extern unsigned int oom_badness(struct task_struct *p, struct mem_cgroup *mem,
+extern unsigned long oom_badness(struct task_struct *p, struct mem_cgroup *mem,
 			const nodemask_t *nodemask, unsigned long totalpages);
 extern void oom_kill_process(struct task_struct *p, gfp_t gfp_mask, int order,
 			     unsigned int points, unsigned long totalpages,
